+++ src/main/java/com/onlinemart/service/impl/VendorServiceImpl.java
@@ -35,18 +35,18 @@ public class VendorServiceImpl implements VendorService {
     }
 
     @Transactional(readOnly = true)
-    public Vendor getVendor(int id) {
+    public Vendor getVendor(Long id) {
         return vendorDao.getVendor(id);
     }
 
     @Override
-    public void deleteVendor(int id) {
+    public void deleteVendor(Long id) {
         vendorDao.deleteVendor(id);
 
     }
 
      @Override
-    public Vendor getVendorById(int id) {
+    public Vendor getVendorById(Long id) {
         return vendorDao.getVendorById(id);
     }
